Pivot to RateLimitController
